third_party/blink/web_tests/fast/table/percent-widths-total-less-than-one.html

<!DOCTYPE HTML>
<style>
    table { border-collapse: collapse; width: 1000px;}
    #green { background-color: green; padding: 0; width: 0.1%;}
    #blue { background-color: blue; padding: 0; width: 0.2%;}
    div { height: 10px; }
</style>
<script src="../../resources/check-layout.js"></script>
<table>
    <td id="green" data-expected-width=333><div></div></td>
    <td id="blue" data-expected-width=667><div></div></td>
</table>
<p> crbug.com/476370: Test cells get appropriate width when total width of cells is less than 1%. </p>
<div id="console"></div>
<script>
    checkLayout('td', document.getElementById('console'));
</script>
